+++ tests/integration/components/power-select-test.js
@@ -70,9 +70,13 @@ const groupedNumbers = [
   "one thousand"
 ];
 
+function typeText(selector, text) {
+  $(selector).val(text);
+  $(selector).trigger('input');
+}
+
 function typeInSearch(text) {
-  $('.ember-power-select-search input, .ember-power-select-trigger-multiple-input').val(text);
-  $('.ember-power-select-search input, .ember-power-select-trigger-multiple-input').trigger('input');
+  typeText('.ember-power-select-search input, .ember-power-select-trigger-multiple-input', text);
 }
 
 /**
@@ -2043,6 +2047,7 @@ test('Passing as options the result of `store.query` works', function(assert) {
 11 - Customization using components
   a) [DONE] selected option can be customized using selectedComponent.
   b) [DONE] the list of options can be customized using optionsComponent.
+  c) [NOT DONE] The selected component receives the search action and can make use of it
 */
 
 moduleForComponent('ember-power-select', 'Integration | Component | Ember Power Select (Customization using components)', {
@@ -2081,6 +2086,23 @@ test('the list of options can be customized using optionsComponent', function(as
   let text = $('.ember-power-select-options').text().trim();
   assert.ok(/Countries:/.test(text), 'The given component is rendered');
   assert.ok(/3\. Russia/.test(text), 'The component has access to the options');
+});
+
+test('The selected component receives the search action', function(assert) {
+  assert.expect(1);
+
+  this.search = (term) => {
+    assert.equal(term, 'foobar', 'The search action received the text typed in the trigger component');
+    return [];
+  };
+
+  this.render(hbs`
+    {{#power-select search=(action search) selected=foo selectedComponent="typeahead-for-test" onchange=(action (mut foo)) as |country|}}
+      {{country.name}}
+    {{/power-select}}
+  `);
+
+  Ember.run(() => typeText('.typeahead-for-test-input', 'foobar'));
 });
 
 /**